requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
PyMuPDF==1.26.3
opencv-python==4.12.0.88
pillow==11.3.0
numpy==2.2.6
pytesseract==0.3.13
pdf2image==1.17.0
torch==2.8.0
transformers==4.55.0
timm==1.0.19
